% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_age_group.R
\name{add_age_group}
\alias{add_age_group}
\title{Add Age Groups and adjusted ages to pd}
\usage{
add_age_group(pd, age_col = "Age")
}
\arguments{
\item{pd}{colData for rse object}

\item{age_col}{name of column containing numeric age information}
}
\value{
colData appended with age_group and adjusted ages
}
\description{
Add Age Groups and adjusted ages to pd
}
\examples{

library(SummarizedExperiment)
pd_test <- colData(rse_bulk_test)
pd_test2 <- add_age_group(pd_test)

}
