% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_categories}
\alias{age_categories}
\title{List of ages defining life stages}
\format{
Named list.
}
\source{
\url{https://github.com/LieberInstitute/DeconvoBuddies/blob/main/data-raw/age_categories.R}
}
\usage{
age_categories
}
\description{
A named list of the minimum and maximum ages for 6 life stages (Prenatal,
Infant, Child, Teen, Adult, and 50+). Used in \code{add_age_group()}.
}
\keyword{datasets}
