% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_ratio2.R
\name{.get_mean_ratio2}
\alias{.get_mean_ratio2}
\title{Get Mean Ratio for Each Gene x Cell Type (Old internal version)}
\usage{
.get_mean_ratio2(
  sce,
  cellType_col = "cellType",
  assay_name = "logcounts",
  add_symbol = TRUE
)
}
\arguments{
\item{sce}{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment-class} object}

\item{cellType_col}{A \code{character(1)} name of the column in the
\link[SummarizedExperiment:SummarizedExperiment-class]{colData()} of \code{sce} that
denotes the cell type or group of interest}

\item{assay_name}{A \code{character(1)} specifying the name of the
\link[SummarizedExperiment:SummarizedExperiment-class]{assay()} in the
\code{sce} object to use to rank expression values. Defaults to \code{logcounts} since
it typically contains the normalized expression values.}

\item{add_symbol}{a logical indicating whether the gene symbol column to the marker stats table}
}
\value{
Table of mean ratio for each x cell type
}
\description{
Calculate the mean ratio value and rank for each gene for each cell type in the \code{sce}
object, to identify effective marker genes for deconvolution.
}
\details{
Improved efficiency and ability to handle large data sets from \code{get_mean_ratio()}.
}
\examples{
#' ## load example SingleCellExperiment
if (!exists("sce_DLPFC_example")) sce_DLPFC_example <- fetch_deconvo_data("sce_DLPFC_example")

## Get the mean ratio for each gene for each cell type defined in `cellType_broad_hc`
.get_mean_ratio2(sce_DLPFC_example, cellType_col = "cellType_broad_hc")

#  gene       cellType.target mean.target cellType    mean ratio rank_ratio anno_ratio
#<chr>      <fct>                 <dbl> <fct>      <dbl> <dbl>      <int> <chr>
# 1 CD22       Oligo                  1.36 OPC       0.0730  18.6          1 Oligo/OPC: 18.625
# 2 LINC01608  Oligo                  2.39 Micro     0.142   16.8          2 Oligo/Micro: 16.829
# 3 FOLH1      Oligo                  1.59 OPC       0.101   15.7          3 Oligo/OPC: 15.684
# 4 SLC5A11    Oligo                  2.14 Micro     0.145   14.7          4 Oligo/Micro: 14.697
# 5 AC012494.1 Oligo                  2.42 OPC       0.169   14.3          5 Oligo/OPC: 14.282

}
\keyword{internal}
