% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_express_fill.R
\name{gene_express_fill}
\alias{gene_express_fill}
\title{Plot the gene expression of a list of genes in a SCE object}
\usage{
gene_express_fill(
  sce,
  genes,
  assay = "logcounts",
  cat = "cellType",
  color_pal = NULL,
  title = NULL
)
}
\arguments{
\item{sce}{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment-class} object}

\item{genes}{A \code{list()} of \code{character(1)} specifying the names of genes to plot,}

\item{assay}{A \code{character(1)} specifying the name of the
\link[SummarizedExperiment:SummarizedExperiment-class]{assay()} in the
\code{sce} object to use to rank expression values. Defaults to \code{logcounts} since
it typically contains the normalized expression values.}

\item{cat}{A \code{character(1)} specifying the name of the categorical variable
to group the cells or nuclei by. Defaults to \code{cellType}.}

\item{color_pal}{A named \code{character(1)} vector that contains a color pallet matching the \code{cat} values}

\item{title}{A \code{character(1)} to title the plot}
}
\value{
A \code{ggplot()} violin plot for selected genes
}
\description{
Plot the gene expression of a list of genes in a SCE object
}
\examples{
rownames(sce.test) <- rowData(sce.test)$Symbol
gene_express_fill(sce = sce.test, genes = c("MBP","SNAP25"))
gene_express_fill(sce = sce.test, genes = c("RNF220","CSF3R","PRDM16","MORN1"))
gene_express_fill(sce = sce.test, genes = c("RNF220"))

}
