% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_ratio2.R
\name{get_mean_ratio2}
\alias{get_mean_ratio2}
\title{Get Mean Ratio for Each Gene x Cell Type}
\usage{
get_mean_ratio2(
  sce,
  cellType_col = "cellType",
  assay_name = "counts",
  add_symbol = FALSE
)
}
\arguments{
\item{sce}{Single cell experiment object}

\item{cellType_col}{Column name on colData of the sce that denotes the celltype}

\item{assay_name}{Name of assay to use for calculation}
}
\value{
Table of mean ratio for each x cell type
}
\description{
Get Mean Ratio for Each Gene x Cell Type
}
\examples{
set.seed(127)
test_sce <- make_test_sce()
mrt <- get_mean_ratio2(test_sce)

}
