% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marker_stats_1vAll}
\alias{marker_stats_1vAll}
\title{1vAll Marker Statistics example data}
\format{
A \code{data.frame} object.
}
\source{
\url{https://github.com/LieberInstitute/DeconvoBuddies/blob/master/inst/extdata/data-raw/RNAScope_prop.R}
}
\usage{
data("marker_stats_1vAll")
}
\description{
A tibble contatinting the marker statistics calculated for 5k genes from
DLPFC snRNA-seq dataset by \code{findMarkers_1vAll}.
}
\details{
3.47 MB
}
\examples{
# To view source
system.file("extdata", "data-raw", "marker_stats_1vAll.R", package = "DeconvoBuddies")
}
\keyword{datasets}
