% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_prop_zero.R
\name{max_prop_zero}
\alias{max_prop_zero}
\title{Find the maximum proportion of zero counts between groups for each gene}
\usage{
max_prop_zero(sce, group_col = "cellType")
}
\arguments{
\item{sce}{Single cell experiment object}

\item{groups}{column names in the colData of sce that define groups of interest}
}
\value{

}
\description{
Find the maximum proportion of zero counts between groups for each gene
}
\examples{
max_prop_zero(sce = sce_ab)
head(max_prop_zero(sce = sce.test))
}
