% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cell_types.R
\name{order_cell_types}
\alias{order_cell_types}
\title{Order Cell Types Glia, then Neurons, then Ambig/drop}
\usage{
order_cell_types(cell_types)
}
\arguments{
\item{cell_types}{}
}
\description{
Order Cell Types Glia, then Neurons, then Ambig/drop
}
