% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_prop_zero.R
\name{prop_zero}
\alias{prop_zero}
\title{Find the poportion of zero counts pre group for each gene}
\usage{
prop_zero(sce, group_col = "cellType")
}
\arguments{
\item{sce}{Single cell experiment object}

\item{groups}{column names in the colData of sce that define groups of interest}
}
\value{

}
\description{
Find the poportion of zero counts pre group for each gene
}
\examples{
prop_zero(sce_ab)
sce_ab$g2 <- paste0(sce_ab$cellType, "_", sce_ab$donor)
prop_zero(sce_ab, group_col = "g2")
prop_zero(sce.test)
}
