% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{pseudobulk}
\alias{pseudobulk}
\title{Pseudobulks sce data over donor x cell type}
\usage{
pseudobulk(
  sce,
  cell_group_cols = c("donor", "cellType"),
  add_symbol = FALSE,
  just_counts = FALSE
)
}
\arguments{
\item{sce}{Single cell experiment object}

\item{cell_group_cols}{list of column names that create pseudobulk groups}

\item{add_symbol}{Use Symbol from rowData as rownames}

\item{just_counts}{Whether to return just the counts matrix or a SummarizedExperiment object (Default)}
}
\value{
Matrix of pseudobukled logcounts or SummarizedExperiment
}
\description{
Pseudobulks sce data over donor x cell type
}
\examples{
## Default, pseudobulk by cellType and donor
pb_label_ab <- pseudobulk(sce_ab, just_counts = FALSE)
head(pb_label_ab)

## Psuedobulk by user defined columns
pb_label_ab_region <- pseudobulk(sce_ab, cell_group_cols = c("cellType", "donor", "region"))
head(pb_label_ab_region)

## Use the Gene Symbol as the rownames
pb_label_test <- pseudobulk(sce.test, cell_group_cols = c("donor", "cellType.Broad"), add_symbol = TRUE)
head(pb_label_test)

## Only output the count matrix
pb_label_test_counts <- pseudobulk(sce.test, cell_group_cols = c("donor", "cellType.Broad"), just_counts = TRUE)
head(pb_label_test_counts)

}
