% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/sce.R
\docType{data}
\name{sce.test}
\alias{sce.test}
\title{Test subset of a real snRNA-seq data}
\format{
A SingleCellExperiment object.

sce with 1000 genes
}
\source{
\url{https://github.com/LieberInstitute/brainstorm/blob/master/data-raw/sce.test.R}
}
\usage{
sce.test
}
\description{
A subset of 1000 genes and 500 nuclei from Tran, Maynard et al., Neuron 2021
including 8 cell types.

Description:
}
\details{
4.26 MB
}
\references{
\enumerate{
\item Pickrell JK, Marioni JC, Pai AA, Degner JF, Engelhardt BE, Nkadori E,
Veyrieras J-B, Stephens M, Gilad Y, Pritchard JK. Understanding mechanisms
underlying human gene expression variation with RNA sequencing. Nature 2010
Apr.
\item Montgomery SB, Sammeth M, Gutierrez-Arcelus M, Lach RP, Ingle C, Nisbett
J, Guigo R, Dermitzakis ET. Transcriptome genetics using second generation
sequencing in a Caucasian population. Nature 2010 Mar.
}
}
\seealso{
\link{get_mean_ratio2}
}
\keyword{datasets}
