% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tableau20}
\alias{tableau20}
\title{A 20 color pallet}
\format{
list of 20 hex colors
}
\source{
\url{https://github.com/LieberInstitute/brainstorm/blob/master/data-raw/tableau20.R}
}
\usage{
tableau20
}
\description{
Adapted from \code{scater}
}
\keyword{datasets}
