% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cell_colors.R
\name{create_cell_colors}
\alias{create_cell_colors}
\title{Create a Cell Color Pallet for Deconvolution}
\usage{
create_cell_colors(
  cell_types = c("Inhib", "Excit", "Astro", "Micro", "Oligo", "OPC"),
  pallet = c("classic", "gg", "tableau"),
  split = NA,
  preview = FALSE
)
}
\arguments{
\item{cell_types}{list of cell types}

\item{pallet}{Choice of base pallet "classic", "gg", or "tableau"}

\item{split}{delineating character in cell names}

\item{preview}{plot preview of colors}
}
\value{

}
\description{
Create a Cell Color Pallet for Deconvolution
}
\examples{
create_cell_colors(pallet = "classic")
create_cell_colors(pallet = "classic", preview = TRUE)
create_cell_colors(pallet = "tableau", preview = TRUE)
create_cell_colors(cell_types = c("A.1","A.2", "B.1", "C"), split = "\\\\.", pallet = "gg",preview = TRUE)
}
