% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeconvoBuddies-package.R
\docType{package}
\name{DeconvoBuddies-package}
\alias{DeconvoBuddies}
\alias{DeconvoBuddies-package}
\title{DeconvoBuddies: Helper Functions for LIBD Deconvolution}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions helpful for LIBD deconvolution project. Includes tools for marker finding with mean ratio, expression plotting, and plotting deconvolution results. Working to include DLPFC datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LieberInstitute/DeconvoBuddies}
  \item Report bugs at \url{https://github.com/LieberInstitute/DeconvoBuddies/issues}
}

}
\author{
\strong{Maintainer}: Louise Huuki-Myers \email{lahuuki@gmail.com} (\href{https://orcid.org/0000-0001-5148-3602}{ORCID})

Other contributors:
\itemize{
  \item Leonardo Collado-Torres \email{lcolladotor@gmail.com} (\href{https://orcid.org/0000-0003-2140-308X}{ORCID}) [contributor]
  \item Nicholas J. Eagles \email{nickeagles77@gmail.com} (\href{https://orcid.org/0000-0002-9808-5254}{ORCID}) [contributor]
}

}
\keyword{internal}
