% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{est_prop}
\alias{est_prop}
\title{Test Estimated Cell Type Proportions}
\format{
A data.frame object.
}
\source{
\url{https://github.com/LieberInstitute/brainstorm/blob/master/data-raw/est_prop.R}
}
\usage{
est_prop
}
\description{
A test dataset of estiamted protions for 5 cell types over 100 samples.
}
\keyword{datasets}
