% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_deconvo_data.R
\name{fetch_deconvo_data}
\alias{fetch_deconvo_data}
\title{Download Human DLPFC Deconvolution Data}
\usage{
fetch_deconvo_data(
  type = c("rse_gene"),
  destdir = tempdir(),
  eh = ExperimentHub::ExperimentHub(),
  bfc = BiocFileCache::BiocFileCache()
)
}
\arguments{
\item{type}{A \code{character(1)} specifying which file you want to download.
\code{rse_gene} {RangedSummarizedExperiment} with 110 bulk RNA-seq samples}

\item{destdir}{The destination directory to where files will be downloaded
to in case the \code{ExperimentHub} resource is not available. If you already
downloaded the files, you can set this to the current path where the files
were previously downloaded to avoid re-downloading them.}

\item{eh}{An \code{ExperimentHub} object
\link[ExperimentHub:ExperimentHub-class]{ExperimentHub-class}.}

\item{bfc}{A \code{BiocFileCache} object
\link[BiocFileCache:BiocFileCache-class]{BiocFileCache-class}. Used when
\code{eh} is not available.}
}
\value{
The requested object
}
\description{
Download Human DLPFC Deconvolution Data
}
\examples{
rse_gene <- fetch_deconvo_data("rse_gene")

}
