% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkers_1vAll.R
\name{findMarkers_1vAll}
\alias{findMarkers_1vAll}
\title{Calculate 1 vs. All standard fold change for each gene x cell type}
\usage{
findMarkers_1vAll(sce, assay = "logcounts", cellType_col = "cellType")
}
\arguments{
\item{sce}{single cell experiment object}

\item{assay}{Name of the assay to use for calculation}

\item{cellType_col}{Column name on colData of the sce that denotes the celltype}
}
\value{
Table of 1 vs. ALL std log fold change + p-values for each gene x cell type
}
\description{
Calculate 1 vs. All standard fold change for each gene x cell type
}
