% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkers_1vAll.R
\name{findMarkers_1vAll}
\alias{findMarkers_1vAll}
\title{Calculate 1 vs. All standard fold change for each gene x cell type, wrapper
function for scran::findMarkers}
\usage{
findMarkers_1vAll(
  sce,
  assay_name = "counts",
  cellType_col = "cellType",
  add_symbol = FALSE,
  mod = "~donor",
  verbose = TRUE
)
}
\arguments{
\item{sce}{single cell experiment object}

\item{assay_name}{Name of the assay to use for calculation}

\item{cellType_col}{Column name on colData of the sce that denotes the celltype}

\item{add_symbol}{Add the gene symbol column to the marker stats table}

\item{mod}{String specifying the model used as design in findMarkers. Can be \code{NULL} if there are no blocking terms with uninteresting factors as documented at \link[scran:pairwiseTTests]{pairwiseTTests}.}

\item{verbose}{Boolean choosing to print progress messages or not}
}
\value{
Table of 1 vs. ALL std log fold change + p-values for each gene x cell type
}
\description{
Calculate 1 vs. All standard fold change for each gene x cell type, wrapper
function for scran::findMarkers
}
\examples{
markers_1vAll <- findMarkers_1vAll(sce.test)
head(markers_1vAll)

}
