% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkers_pw.R
\name{findMarkers_pw}
\alias{findMarkers_pw}
\title{Calculate pw standard fold change for each gene x cell type}
\usage{
findMarkers_pw(
  sce,
  assay_name = "counts",
  cellType_col = "cellType",
  add_symbol = FALSE
)
}
\arguments{
\item{sce}{single cell experiment object}

\item{assay_name}{Name of the assay to use for calculation}

\item{cellType_col}{Column name on colData of the sce that denotes the celltype}

\item{add_symbol}{Add the gene symbol column to the marker stats table}
}
\value{
Table of 1 vs. ALL std log fold change + p-values for each gene x cell type
}
\description{
Calculate pw standard fold change for each gene x cell type
}
\examples{
TODO
}
