% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_ratio.R
\name{get_mean_ratio}
\alias{get_mean_ratio}
\title{Get Mean Ratio for Each Gene x Cell Type}
\usage{
get_mean_ratio(
  sce,
  cellType_col = "cellType",
  assay_name = "logcounts",
  gene_ensembl = NULL,
  gene_name = NULL
)
}
\arguments{
\item{sce}{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment-class} object}

\item{cellType_col}{A \code{character(1)} name of the column in the
\link[SummarizedExperiment:SummarizedExperiment-class]{colData()} of \code{sce} that
denotes the cell type or group of interest}

\item{assay_name}{A \code{character(1)} specifying the name of the
\link[SummarizedExperiment:SummarizedExperiment-class]{assay()} in the
\code{sce} object to use to rank expression values. Defaults to \code{logcounts} since
it typically contains the normalized expression values.}

\item{gene_ensembl}{A \code{character(1)} specifying the \code{rowData(sce_pseudo)}
column with the ENSEMBL gene IDs. This will be used by \code{layer_stat_cor()}.}

\item{gene_name}{A \code{character(1)} specifying the \code{rowData(sce_pseudo)}
column with the gene names (symbols).}
}
\value{
Table of mean ratio for each x cell type
}
\description{
Calculate the mean ratio value and rank for each gene for each cell type in the \code{sce}
object, to identify effective marker genes for deconvolution.
}
\details{
Improved efficiency and ability to handle large data sets from \code{get_mean_ratio()}.
}
\examples{
get_mean_ratio(sce.test)

#Specify gene_name as the "Symbol" column from rowData
# this will be added to the marker stats output
get_mean_ratio(sce.test, gene_name = "Symbol")

}
