% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_ratio.R
\name{get_mean_ratio}
\alias{get_mean_ratio}
\title{Get Mean Ratio for Each Gene x Cell Type}
\usage{
get_mean_ratio(
  sce,
  cellType_col,
  assay_name = "logcounts",
  gene_ensembl = NULL,
  gene_name = NULL
)
}
\arguments{
\item{sce}{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment-class} object}

\item{cellType_col}{A \code{character(1)} name of the column in the
\link[SummarizedExperiment:SummarizedExperiment-class]{colData()} of \code{sce} that
denotes the cell type or group of interest}

\item{assay_name}{A \code{character(1)} specifying the name of the
\link[SummarizedExperiment:SummarizedExperiment-class]{assay()} in the
\code{sce} object to use to rank expression values. Defaults to \code{logcounts} since
it typically contains the normalized expression values.}

\item{gene_ensembl}{A \code{character(1)} specifying the \code{rowData(sce_pseudo)}
column with the ENSEMBL gene IDs. This will be used by \code{layer_stat_cor()}.}

\item{gene_name}{A \code{character(1)} specifying the \code{rowData(sce_pseudo)}
column with the gene names (symbols).}
}
\value{
A \code{tibble} with the \code{MeanRatio} values for each gene x cell type.
\itemize{
\item \code{gene} is the name of the gene (from rownames(\code{sce})).
\item \code{cellType.target} is the cell type we're finding marker genes for.
\item \code{mean.target} is the mean expression of \code{gene} for \code{cellType.target}.
\item \code{cellType.2nd} is the second highest non-target cell type.
\item \code{mean.2nd} is the mean expression of \code{gene} for \code{cellType.2nd}.
\item \code{MeanRatio} is the ratio of \code{mean.target/mean.2nd}.
\item \code{MeanRatio.rank} is the rank of \code{MeanRatio} for the cell type.
\item \code{MeanRatio.anno} is an annotation of the \code{MeanRatio} calculation helpful for plotting.
\item \code{gene_ensembl} & \code{gene_name} optional cols spcified by the user to add gene infomation
}
}
\description{
Calculate the mean ratio value and rank for each gene for each cell type in the \code{sce}
object, to identify effective marker genes for deconvolution.
}
\details{
Improved argument names and documentaion, but same functionalty from \code{get_mean_ratio2()}.
}
\examples{
## Get the mean ratio for each gene for each cell type defined in `cellType`
get_mean_ratio(sce.test, cellType_col = "cellType")

#   gene            cellType.target mean.target cellType.2nd mean.2nd MeanRatio MeanRatio.rank MeanRatio.anno        
#  <chr>           <fct>                 <dbl> <fct>           <dbl>     <dbl>          <int> <chr>                 
# 1 ENSG00000230615 Inhib.2               1.00  Excit.2         0.239      4.20              1 Inhib.2/Excit.2: 4.197
# 2 ENSG00000162512 Inhib.2               1.71  Astro           0.512      3.35              2 Inhib.2/Astro: 3.346  
# 3 ENSG00000137965 Inhib.2               0.950 Astro           0.413      2.30              3 Inhib.2/Astro: 2.301  
# 4 ENSG00000060718 Inhib.2               3.32  Astro           1.47       2.26              4 Inhib.2/Astro: 2.264  
# 5 ENSG00000162631 Inhib.2               3.55  Astro           1.62       2.19              5 Inhib.2/Astro: 2.19  

# Option to specify gene_name as the "Symbol" column from rowData
# this will be added to the marker stats output
SummarizedExperiment::rowData(sce.test)
get_mean_ratio(sce.test, cellType_col = "cellType", gene_name = "Symbol", gene_ensembl = "gene_id")
# A tibble: 1,778 × 10
# gene            cellType.target mean.target cellType.2nd mean.2nd MeanRatio MeanRatio.rank MeanRatio.anno   gene_ensembl gene_name
# <chr>           <fct>                 <dbl> <fct>           <dbl>     <dbl>          <int> <chr>            <chr>        <chr>    
# 1 ENSG00000230615 Inhib.2               1.00  Excit.2         0.239      4.20              1 Inhib.2/Excit.2… ENSG0000023… AL139220…
# 2 ENSG00000162512 Inhib.2               1.71  Astro           0.512      3.35              2 Inhib.2/Astro: … ENSG0000016… SDC3     
# 3 ENSG00000137965 Inhib.2               0.950 Astro           0.413      2.30              3 Inhib.2/Astro: … ENSG0000013… IFI44    

}
