% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_ratio.R
\name{get_mean_ratio}
\alias{get_mean_ratio}
\title{Get Mean Ratio for Each Gene x Cell Type}
\usage{
get_mean_ratio(
  sce,
  cellType_col,
  assay_name = "logcounts",
  gene_ensembl = NULL,
  gene_name = NULL
)
}
\arguments{
\item{sce}{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment-class} object}

\item{cellType_col}{A \code{character(1)} name of the column in the
\link[SummarizedExperiment:SummarizedExperiment-class]{colData()} of \code{sce} that
denotes the cell type or group of interest}

\item{assay_name}{A \code{character(1)} specifying the name of the
\link[SummarizedExperiment:SummarizedExperiment-class]{assay()} in the
\code{sce} object to use to rank expression values. Defaults to \code{logcounts} since
it typically contains the normalized expression values.}

\item{gene_ensembl}{A \code{character(1)} specifying the \code{rowData(sce_pseudo)}
column with the ENSEMBL gene IDs. This will be used by \code{layer_stat_cor()}.}

\item{gene_name}{A \code{character(1)} specifying the \code{rowData(sce_pseudo)}
column with the gene names (symbols).}
}
\value{
A \code{tibble} with the \code{MeanRatio} values for each gene x cell type.
\itemize{
\item \code{gene} is the name of the gene (from rownames(\code{sce})).
\item \code{cellType.target} is the cell type we're finding marker genes for.
\item \code{mean.target} is the mean expression of \code{gene} for \code{cellType.target}.
\item \code{cellType.2nd} is the second highest non-target cell type.
\item \code{mean.2nd} is the mean expression of \code{gene} for \code{cellType.2nd}.
\item \code{MeanRatio} is the ratio of \code{mean.target/mean.2nd}.
\item \code{MeanRatio.rank} is the rank of \code{MeanRatio} for the cell type.
\item \code{MeanRatio.anno} is an annotation of the \code{MeanRatio} calculation helpful for plotting.
\item \code{gene_ensembl} & \code{gene_name} optional cols spcified by the user to add gene infomation
}
}
\description{
Calculate the mean ratio value and rank for each gene for each cell type in the \code{sce}
object, to identify effective marker genes for deconvolution.
}
\details{
Improved argument names and documentaion, but same functionalty from \code{get_mean_ratio2()}.
}
\examples{
## load example SingleCellExperiment (4.79 MB)
if (!exists("sce_DLPFC_example")) sce_DLPFC_example <- fetch_deconvo_data("sce_DLPFC_example")

## Get the mean ratio for each gene for each cell type defined in `cellType_broad_hc`
get_mean_ratio(sce_DLPFC_example, cellType_col = "cellType_broad_hc")

# A tibble: 764 × 8
#gene     cellType.target mean.target cellType.2nd mean.2nd MeanRatio MeanRatio.rank MeanRatio.anno         
#<chr>    <fct>                 <dbl> <fct>           <dbl>     <dbl>          <int> <chr>                  
#  1 OBI1-AS1 Astro                  3.74 Oligo           0.291     12.9               1 Astro/Oligo: 12.858    
#2 PRDM16   Astro                  1.97 EndoMural       0.175     11.3               2 Astro/EndoMural: 11.254
#3 PAMR1    Astro                  2.21 Excit           0.208     10.7               3 Astro/Excit: 10.65     
#4 RGS20    Astro                  1.71 Inhib           0.173      9.88              4 Astro/Inhib: 9.876     
#5 HPSE2    Astro                  2.49 Inhib           0.282      8.82              5 Astro/Inhib: 8.823   

# Option to specify gene_name as the "Symbol" column from rowData
# this will be added to the marker stats output
SummarizedExperiment::rowData(sce_DLPFC_example)
#  gene     cellType.target mean.target cellType.2nd mean.2nd MeanRatio MeanRatio.rank MeanRatio.anno         
#<chr>    <fct>                 <dbl> <fct>           <dbl>     <dbl>          <int> <chr>                  
#1 OBI1-AS1 Astro                  3.74 Oligo           0.291     12.9               1 Astro/Oligo: 12.858    
#2 PRDM16   Astro                  1.97 EndoMural       0.175     11.3               2 Astro/EndoMural: 11.254
#3 PAMR1    Astro                  2.21 Excit           0.208     10.7               3 Astro/Excit: 10.65     
#4 RGS20    Astro                  1.71 Inhib           0.173      9.88              4 Astro/Inhib: 9.876     
#5 HPSE2    Astro                  2.49 Inhib           0.282      8.82              5 Astro/Inhib: 8.823   

get_mean_ratio(sce_DLPFC_example, cellType_col = "cellType_broad_hc", gene_name = "gene_name", gene_ensembl = "gene_id")
# A tibble: 764 × 10
# gene            cellType.target mean.target cellType.2nd mean.2nd MeanRatio MeanRatio.rank MeanRatio.anno   gene_ensembl gene_name
# <chr>           <fct>                 <dbl> <fct>           <dbl>     <dbl>          <int> <chr>            <chr>        <chr>    
# 1 OBI1-AS1 Astro                  3.74 Oligo           0.291     12.9               1 Astro/Oligo: 12.858     ENSG00000234… OBI1-AS1 
# 2 PRDM16   Astro                  1.97 EndoMural       0.175     11.3               2 Astro/EndoMural: 11.254 ENSG00000142… PRDM16   
# 3 PAMR1    Astro                  2.21 Excit           0.208     10.7               3 Astro/Excit: 10.65      ENSG00000149… PAMR1    

}
