% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_ratio.R
\name{get_mean_ratio}
\alias{get_mean_ratio}
\title{Get Mean Ratio for Each Gene x Cell Type}
\usage{
get_mean_ratio(sce, cellType_col = "cellType")
}
\arguments{
\item{sce}{Single cell experiment object}

\item{cellType_col}{Column name on colData of the sce that denotes the celltype}
}
\value{
Table of mean ratio for each x cell type
}
\description{
Get Mean Ratio for Each Gene x Cell Type
}
\examples{
set.seed(127)
test_sce <- make_test_sce()
get_mean_ratio(test_sce)

}
