% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_test_sce.R
\name{make_test_sce}
\alias{make_test_sce}
\title{Simulate a SingleCellExperiment for testing}
\usage{
make_test_sce(n_cell = 100, n_gene = 100, n_cellType = 4, n_donor = 2)
}
\arguments{
\item{n_cell}{Number of cells}

\item{n_gene}{Number of genes}

\item{n_cellType}{Number of cell types}

\item{n_donor}{Number of donors}
}
\value{
SingelCellExperiment object with randomly generated counts and colData
}
\description{
Simulate a SingleCellExperiment for testing
}
\examples{
test <- make_test_sce()
table(test$cellType, test$donor)
}
