% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marker_test}
\alias{marker_test}
\title{Markers stats from sce_DLPFC_example}
\format{
A \code{tibble::tibble()}. See \code{get_mean_ratio()} for more details on the column
names.
}
\source{
\url{https://github.com/LieberInstitute/DeconvoBuddies/blob/master/inst/extdata/data-raw/marker_test.R}
}
\usage{
data("marker_test")
}
\description{
A tibble containing the marker stats from \code{get_mean_ratio()} for
\code{sce_DLPFC_example}.
}
\details{
402.60 kB
}
\examples{
# To view source
system.file("extdata", "data-raw", "marker_test.R", package = "DeconvoBuddies")
}
\keyword{datasets}
