% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marker_test}
\alias{marker_test}
\title{Markers stats from sce_DLPFC_example}
\format{
tibble.
}
\source{
\url{https://github.com/LieberInstitute/brainstorm/blob/master/data-raw/marker_test.R}
}
\usage{
marker_test
}
\description{
A tibble containing the marker stats from \code{get_mean_ratio} for \code{sce_DLPFC_example}.
}
\details{
402.60 kB
}
\keyword{datasets}
