% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cell_types.R
\name{order_cell_types}
\alias{order_cell_types}
\title{Order Cell Types Glia, then Neurons, then Ambig/drop}
\usage{
order_cell_types(cell_types)
}
\arguments{
\item{cell_types}{A \code{list()} of \code{character(1)} specifying the names of cell types}
}
\value{
A reordered \code{list()} with Glia cell types before neurons
}
\description{
Order Cell Types Glia, then Neurons, then Ambig/drop
}
\examples{

order_cell_types(c("Astro", "Excit", "Inhib", "Oligo"))

}
