% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_composition_bar.R
\name{plot_composition_bar}
\alias{plot_composition_bar}
\title{Create barplot of average cell type composition}
\usage{
plot_composition_bar(
  prop_long,
  sample_col = "RNum",
  x_col = "ALL",
  prop_col = "prop",
  ct_col = "cell_type",
  add_text = TRUE,
  min_prop_text = 0
)
}
\arguments{
\item{prop_long}{data.frame of cell type portions in long form}

\item{sample_col}{name of column that identifies samples}

\item{x_col}{category to divide samples by}

\item{prop_col}{name of column containing proportion values}

\item{ct_col}{name of column containing cell type names}

\item{add_text}{Add rounded proportion value to bars}

\item{min_prop_text}{Minimum proportion to display text}
}
\value{
stacked barplot representing mean proportion of cell types for each group
}
\description{
Create barplot of average cell type composition
}
\examples{
pd <- SummarizedExperiment::colData(rse_bulk_test) \%>\%
    as.data.frame()

est_prop_long <- est_prop \%>\%
    tibble::rownames_to_column("RNum") \%>\%
    tidyr::pivot_longer(!RNum, names_to = "cell_type", values_to = "prop") \%>\%
    dplyr::left_join(pd \%>\% dplyr::select(RNum, Dx)) \%>\%
    dplyr::mutate(a = "a")

plot_composition_bar(est_prop_long)
plot_composition_bar(est_prop_long, x_col = "Dx")
plot_composition_bar(est_prop_long, x_col = "Dx", min_prop_text = 0.1)
plot_composition_bar(est_prop_long, x_col = "RNum", add_text = FALSE)
}
