% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_express.R
\name{plot_gene_express}
\alias{plot_gene_express}
\title{Plot the gene expression of a list of genes in a SCE object}
\usage{
plot_gene_express(
  sce,
  genes,
  assay_name = "logcounts",
  cat = "cellType",
  color_pal = NULL,
  title = NULL,
  plot_points = FALSE
)
}
\arguments{
\item{sce}{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment-class} object}

\item{genes}{A \code{list()} of \code{character(1)} specifying the genes to plot, this should match the format of \code{rownames(sce)}}

\item{assay_name}{A \code{character(1)} specifying the name of the
\link[SummarizedExperiment:SummarizedExperiment-class]{assay()} in the
\code{sce} object to use to rank expression values. Defaults to \code{logcounts} since
it typically contains the normalized expression values.}

\item{cat}{A \code{character(1)} specifying the name of the categorical variable
to group the cells or nuclei by. Defaults to \code{cellType}.}

\item{color_pal}{A named \code{character(1)} vector that contains a color pallet matching the \code{cat} values.}

\item{title}{A \code{character(1)} to title the plot}

\item{plot_points}{A logical indicating whether to plot points over the violin,
defaults to \code{FALSE} as these often become overplotted and quite large (especially when saved as PDF)}
}
\value{
A \code{ggplot()} violin plot for selected genes
}
\description{
This function plots the expression of one or more genes as a violin plot,
over a user defined category, typically a cell type annotation.
}
\examples{
## Using Symbol as rownames makes this more human readable
rownames(sce.test) <- rowData(sce.test)$Symbol
plot_gene_express(sce = sce.test, genes = c("F3"))
plot_gene_express(sce = sce.test, genes = c("RNF220","CSF3R"))
plot_gene_express(sce = sce.test, genes = c("RNF220","CSF3R"), points = TRUE)
plot_gene_express(sce = sce.test, assay_name = "counts", genes = c("RNF220","CSF3R"))
plot_gene_express(sce = sce.test, assay_name = "counts", genes = c("RNF220","CSF3R"), title = "Inhib Markers")

}
\seealso{
Other expression plotting functions: 
\code{\link{plot_marker_express_ALL}()},
\code{\link{plot_marker_express}()}
}
\concept{expression plotting functions}
