% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_marker_express.R
\name{plot_marker_express}
\alias{plot_marker_express}
\title{create plotExpression plots for top n genes for each cell type}
\usage{
plot_marker_express(
  sce,
  stats,
  cell_type,
  n_genes,
  rank_col,
  anno_col,
  cellType_col = "cellType"
)
}
\arguments{
\item{sce}{sce object that contains logcounts assay}

\item{stats}{table generated by get_mean_ratio and/or findMarkers_1vAll}

\item{cell_type}{taret cel type for markers}

\item{n_genes}{number of markers you'd like to plot}

\item{rank_col}{name of column to rank genes by}

\item{anno_col}{name of column containing annotation}

\item{cellType_col}{came of colData column containing cell type for sce data}
}
\value{
plotExpression style violin plot for selected marker genes
}
\description{
create plotExpression plots for top n genes for each cell type
}
\examples{

}
