% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_marker_express.R
\name{plot_marker_express}
\alias{plot_marker_express}
\title{Plot gene expression violin plots for top marker genes for one cell type}
\usage{
plot_marker_express(
  sce,
  stats,
  cell_type,
  n_genes = 4,
  rank_col = "rank_ratio",
  anno_col = "anno_ratio",
  cellType_col = "cellType",
  color_pal = NULL,
  plot_points = FALSE
)
}
\arguments{
\item{sce}{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment-class} object}

\item{stats}{A \code{data.frame()} generated by get_mean_ratio and/or findMarkers_1vAll}

\item{cell_type}{A \code{character()} target cell type to plot markers for}

\item{n_genes}{An \code{integer()} of number of markers you'd like to plot}

\item{rank_col}{The \code{character()} name of column to rank genes by in \code{stats}}

\item{anno_col}{The \code{character()} name of column containing annotation in \code{stats}}

\item{cellType_col}{The \code{character()} name of colData column containing cell type for sce data,
matches \code{cellType.target} in \code{stats}}
}
\value{
plotExpression style violin plot for selected marker genes
}
\description{
This function plots the top n marker genes for a specified cell type based off of
the \code{stats} table from \code{get_mean_ratio2()}.
The gene expression is plotted as violin plot with \code{plot_gene_express} and adds
annotations to each plot.
}
\examples{
plot_marker_express(sce = sce.test, stat = marker_test, cell_type = "Astro")
plot_marker_express(sce = sce.test, stat = marker_test, cell_type = "Micro", n_genes = 1, rank_col = "rank_ratio", anno_col = "anno_ratio", plot_points = TRUE)
plot_marker_express(sce = sce.test, stat = marker_test, cell_type = "Oligo", n_genes = 10, rank_col = "rank_ratio", anno_col = "anno_ratio")
plot_marker_express(sce = sce.test, stat = marker_test, cell_type = "Excit.1", n_genes = 5, rank_col = "rank_ratio", anno_col = "anno_ratio")
}
\seealso{
Other expression plotting functions: 
\code{\link{plot_gene_express}()},
\code{\link{plot_marker_express_ALL}()},
\code{\link{plot_marker_express_List}()}
}
\concept{expression plotting functions}
