% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_marker_express_List.R
\name{plot_marker_express_List}
\alias{plot_marker_express_List}
\title{Plot the nested-listed genes as a multi-page pdf}
\usage{
plot_marker_express_List(
  sce,
  stats,
  pdf_fn = "marker_expression.pdf",
  gene_list,
  cellType_col = "cellType",
  gene_name_col = "Symbol",
  color_pal = NULL,
  plot_points = FALSE
)
}
\arguments{
\item{sce}{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment-class} object}

\item{pdf_fn}{A \code{character()} of the pdf filename to plot to}

\item{gene_list}{A named list contating the names of genes to plot}

\item{cellType_col}{The \code{character()} name of colData column containing cell type for sce data}

\item{gene_name_col}{The \code{character()} name of rowData matching the gene name from \code{gene_list}}
}
\value{
A pdf with violin plots for the expression of top marker genes for all cell types
}
\description{
This function plots a nested list of genes as a multipage PDF, one for each sub list.
A use case is plotting known marker genes for multiple cell types over cell type
clusters with unknown identities.
}
\examples{
marker_list <- list(Oligo = c("RNF220","CLIC4"), Astro = c("PRDM16","F3"))

plot_marker_express_List(sce.test, gene_list = marker_list, pdf_fn = "./plots/test_marker_expression_List.pdf")
}
\seealso{
Other expression plotting functions: 
\code{\link{plot_gene_express}()},
\code{\link{plot_marker_express_ALL}()},
\code{\link{plot_marker_express}()}
}
\concept{expression plotting functions}
