% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_marker_express_List.R
\name{plot_marker_express_List}
\alias{plot_marker_express_List}
\title{Plot a nested list of genes as a multi-page pdf}
\usage{
plot_marker_express_List(
  sce,
  gene_list,
  pdf_fn = "marker_expression.pdf",
  cellType_col = "cellType",
  gene_name_col = "Symbol",
  color_pal = NULL,
  plot_points = FALSE
)
}
\arguments{
\item{sce}{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment-class} object}

\item{gene_list}{A named list containing the names of genes to plot}

\item{pdf_fn}{A \code{character()} of the pdf file name to plot to}

\item{cellType_col}{The \code{character()} name of \code{colData} column containing cell type for \code{sce} data}

\item{gene_name_col}{The \code{character()} name of \code{rowData} matching the gene name from \code{gene_list}}

\item{color_pal}{A named \code{character(1)} vector that contains a color pallet matching the values in \code{cellType_col}.}

\item{plot_points}{A logical indicating whether to plot points over the violin,
defaults to \code{FALSE} as these often become over plotted and quite large (especially when saved as PDF)}
}
\value{
A pdf with violin plots for the expression of top marker genes for all cell types
}
\description{
This function plots a nested list of genes as a multi-page PDF, one for each sub list.
A use case is plotting known marker genes for multiple cell types over cell type
clusters with unknown identities.
}
\examples{
marker_list <- list(Oligo = c("RNF220", "CLIC4"), Astro = c("PRDM16", "F3"))
# plot_marker_express_List(sce.test, gene_list = marker_list, pdf_fn = "./plots/test_marker_expression_List.pdf")
# plot_marker_express_List(sce.test, gene_list = marker_list, cellType_col = "cellType.Broad", pdf_fn = "./plots/test_marker_expression_List.pdf")
# plot_marker_express_List(sce.test, gene_list = marker_list, pdf_fn = "./plots/test_marker_expression_List.pdf", plot_points = TRUE)

}
\seealso{
Other expression plotting functions: 
\code{\link{plot_gene_express}()},
\code{\link{plot_marker_express_ALL}()},
\code{\link{plot_marker_express}()}
}
\concept{expression plotting functions}
