% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{pseudobulk}
\alias{pseudobulk}
\title{Pseudobulks sce data over donor x cell type}
\usage{
pseudobulk(sce, cellType_col)
}
\arguments{
\item{sce}{Single cell experiment object}

\item{cellType_col}{Column name on colData of the sce that denotes the celltype}
}
\value{

}
\description{
Pseudobulks sce data over donor x cell type
}
