% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{pseudobulk}
\alias{pseudobulk}
\title{Pseudobulks sce data over donor x cell type}
\usage{
pseudobulk(sce, cell_group_cols = c("donor", "cellType"), add_symbol = FALSE)
}
\arguments{
\item{sce}{Single cell experiment object}

\item{cell_group_cols}{list of column names that create pseudobulk groups}

\item{add_symbol}{Use Symbol from rowData as rownames}
}
\value{

}
\description{
Pseudobulks sce data over donor x cell type
}
\examples{
pb_ab <- pseudobulk(sce_ab)
head(pb_ab)

pb_ab_region <- pseudobulk(sce_ab, cell_group_cols = c("cellType","donor", "region"))
head(pb_ab_region)

pb_test <- pseudobulk(sce.test, cell_group_cols = c("donor","cellType.Broad"))
head(pb_test)
}
