% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNAScope_prop}
\alias{RNAScope_prop}
\title{Cell Type Proportions estimated from RNAScope}
\format{
A data.frame object.
}
\source{
\url{https://github.com/LieberInstitute/brainstorm/blob/master/data-raw/RNAScope.R}
}
\usage{
RNAScope_prop
}
\description{
Cell type proportion estimates from high quality images from
Huuki-Myers et al., bioRxiv, 2024,
doi: \url{https://doi.org/10.1101/2024.02.09.579665}.
}
\details{
11.49 kB

These are the columns of the \code{data.frame} object:
\itemize{
\item \code{SAMPLE_ID} : DLPFC Tissue block + RNAScope combination.
\item \code{Sample} : DLFPC Tissue block (Donor BrNum + DLPFC position).
\item \code{Combo} : RNAScope probe combination, either "Circle" marking cell types Astro
Endo, Inhib, or "Star" marking Excit, Micro, and OligoOPC.
\item \code{cell_type} : The cell type measured.
\item \code{Confidence} : Image confidence, this dataset has been filtered to the high & Ok confidence images.
\item \code{n_cell} : the number of cells counted for the Sample and cell type.
\item \code{prop} : the calculated cell type proportion from n_cell.
\item \code{n_cell_sn} : number of nuclei in the corresponding snRNA-seq data.
\item \code{prop_sn} : cell type proportion from the snRNA-seq data.
}
}
\keyword{datasets}
