% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNAScope_prop}
\alias{RNAScope_prop}
\title{Cell Type Proportions estimated from RNAScope}
\format{
A data.frame object.
}
\source{
\url{https://github.com/LieberInstitute/brainstorm/blob/master/data-raw/RNAScope.R}
}
\usage{
RNAScope_prop
}
\description{
Cell type proportion estimates from high quality images from Deconvolution
Benchmark (TODO cite paper)
}
\details{
\code{SAMPLE_ID} : DLPFC Tissue block + RNAScope combination.
\code{Sample} : DLFPC Tissue block (Donor BrNum + DLPFC position).
\code{Combo} : RNAScope probe combination, either "Circle" marking cell types Astro
Endo, Inhib, or "Star" marking Excit, Micro, and OligoOPC.
\code{cell_type} : The cell type measured.
\code{Confidence} : Image confidence, this dataset has been filtered to the high & Ok confidence images.
\code{n_cell} : the number of cells counted for the Sample and cell type.
\code{prop} : the calculated cell type proportion from n_cell.
\code{n_cell_sn} : number of nuclei in the corresponding snRNA-seq data.
\code{prop_sn} : cell type proportion from the snRNA-seq data.
}
\keyword{datasets}
