% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rse_bulk_test}
\alias{rse_bulk_test}
\title{Test bulk rse dataset}
\format{
A \link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment} object.
}
\source{
\url{https://github.com/LieberInstitute/DeconvoBuddies/blob/master/inst/extdata/data-raw/rse_bulk_test.R}
}
\usage{
data("rse_bulk_test")
}
\description{
A test \code{rse_gene} object with data for 1000 genes across 100 samples.
}
\details{
976.77 kB
}
\examples{
# To view source
system.file("extdata", "data-raw", "rse_bulk_test.R", package = "DeconvoBuddies")
}
\keyword{datasets}
