% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rse_bulk_test}
\alias{rse_bulk_test}
\title{Test bulk rse dataset}
\format{
A SummarizedExperiment object.
}
\source{
\url{https://github.com/LieberInstitute/brainstorm/blob/master/data-raw/genotyped.R}
}
\usage{
rse_bulk_test
}
\description{
TODO Description.
}
\keyword{datasets}
