% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rse_bulk_test}
\alias{rse_bulk_test}
\title{Test bulk rse dataset}
\format{
A SummarizedExperiment object.
}
\source{
\url{https://github.com/LieberInstitute/brainstorm/blob/master/data-raw/genotyped.R}
}
\usage{
rse_bulk_test
}
\description{
A test rse_gene object, 1000 genes x 100 samples
}
\details{
976.77 kB
}
\keyword{datasets}
