% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\docType{data}
\name{sce.test}
\alias{sce.test}
\title{Genome samples processed data}
\format{
sce with 1000 genes
}
\description{
Description:
}
\references{
\enumerate{
\item Pickrell JK, Marioni JC, Pai AA, Degner JF, Engelhardt BE, Nkadori E,
Veyrieras J-B, Stephens M, Gilad Y, Pritchard JK. Understanding mechanisms
underlying human gene expression variation with RNA sequencing. Nature 2010
Apr.
\item Montgomery SB, Sammeth M, Gutierrez-Arcelus M, Lach RP, Ingle C, Nisbett
J, Guigo R, Dermitzakis ET. Transcriptome genetics using second generation
sequencing in a Caucasian population. Nature 2010 Mar.
}
}
\seealso{
\link{get_mean_ratio2}
}
\keyword{datasets}
