% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_ab}
\alias{sce_ab}
\title{Toy SCE object for testing}
\format{
A \link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} object.
}
\source{
\url{https://github.com/LieberInstitute/DeconvoBuddies/blob/master/inst/extdata/data-raw/sce_ab.R}
}
\usage{
data("sce_ab")
}
\description{
An example \code{sce} object for testing, with two cell types A and B.
}
\details{
Generated with \code{DeconvoBuddies::make_test_sce()}
38.26 kB
}
\examples{
# To view source
system.file("extdata", "data-raw", "sce_ab.R", package = "DeconvoBuddies")
}
\keyword{datasets}
