% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_ab}
\alias{sce_ab}
\title{Toy SCE object for testing}
\format{
A SingleCellExperiment object.
}
\source{
\url{https://github.com/LieberInstitute/brainstorm/blob/master/data-raw/sce_ab.R}
}
\usage{
sce_ab
}
\description{
TODO Description.
}
\keyword{datasets}
