% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_ab}
\alias{sce_ab}
\title{Toy SCE object for testing}
\format{
A SingleCellExperiment object.
}
\source{
\url{https://github.com/LieberInstitute/brainstorm/blob/master/data-raw/sce_ab.R}
}
\usage{
sce_ab
}
\description{
An example sce object for testing, with two cell types A and B.
}
\details{
Generated with DeconvoBuddies::make_test_sce()
38.26 kB
}
\keyword{datasets}
